norm: co-locate normalizers, globalize code/extension cleanup
Address review feedback on the norm migration:

- Move each normalizer function next to the model it operates on (bill, org,
  pay, tax), keeping only the norm.Register manifest in each package's norm.go,
  mirroring how the rules package keeps xxxRules() in the model files.

- Globalize the two uniform cleanups as ordinary (single-tier) normalizers so
  individual normalizers no longer repeat them:
  - tax.Extensions: cleaned everywhere via a registered norm.For.
  - cbc.Code: trimmed everywhere via a registered norm.For.

- Redesign cbc.Code normalization/validation so the global code normalizer is
  safe and uniform (no special "final pass" tier needed):
  - NormalizeCode now only trims surrounding whitespace.
  - NormalizeStrictCode carries the previous canonical cleaning (collapse
    separators, strip invalid chars); NormalizeUpper/Alphanumerical/Numerical
    build on it, preserving their behavior.
  - Default Code validation now rejects only leading/trailing whitespace
    (CodePatternLenient); the strict pattern is exposed as the cbc.StrictCode
    validator, applied where strictness is required (e.g. tax.Identity scheme).

- Remove the now-redundant no-argument Normalize() methods on the simple org
  and pay value types; the registered normalizer functions plus the global
  cleaners cover them. Callers updated to norm.Normalize.

Full suite green; data regenerated (cbc/tax rules + cbc code schema only).
