+++ cbc/code_test.go
@@ -84,9 +84,9 @@ func TestNormalizeCode(t *testing.T) {
 			want: cbc.Code("foo-bar1"),
 		},
 		{
-			name: "invalid chars",
+			name: "invalid chars left untouched (lenient default)",
 			code: cbc.Code("f$oo-bar1!"),
-			want: cbc.Code("foo-bar1"),
+			want: cbc.Code("f$oo-bar1!"),
 		},
 		{
 			name: "multiple spaces",
@@ -94,14 +94,14 @@ func TestNormalizeCode(t *testing.T) {
 			want: cbc.Code("foo bar dome"),
 		},
 		{
-			name: "multiple symbols 1",
+			name: "internal symbols left untouched (lenient default)",
 			code: cbc.Code("foo- bar-$dome"),
-			want: cbc.Code("foo-bar-dome"),
+			want: cbc.Code("foo- bar-$dome"),
 		},
 		{
-			name: "multiple symbols 2",
+			name: "internal repeats left untouched (lenient default)",
 			code: cbc.Code("FOO  BAR--DOME"),
-			want: cbc.Code("FOO BAR-DOME"),
+			want: cbc.Code("FOO  BAR--DOME"),
 		},
 		{
 			name: "colons",
@@ -117,6 +117,45 @@ func TestNormalizeCode(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			assert.Equal(t, tt.want, cbc.NormalizeCode(tt.code))
+		})
+	}
+}
+
+func TestNormalizeStrictCode(t *testing.T) {
+	tests := []struct {
+		name string
+		code cbc.Code
+		want cbc.Code
+	}{
+		{
+			name: "whitespace",
+			code: cbc.Code(" foo-bar1  "),
+			want: cbc.Code("foo-bar1"),
+		},
+		{
+			name: "invalid chars",
+			code: cbc.Code("f$oo-bar1!"),
+			want: cbc.Code("foo-bar1"),
+		},
+		{
+			name: "multiple symbols 1",
+			code: cbc.Code("foo- bar-$dome"),
+			want: cbc.Code("foo-bar-dome"),
+		},
+		{
+			name: "multiple symbols 2",
+			code: cbc.Code("FOO  BAR--DOME"),
+			want: cbc.Code("FOO BAR-DOME"),
+		},
+		{
+			name: "colons",
+			code: cbc.Code("0088:1234567891234"),
+			want: cbc.Code("0088:1234567891234"),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			assert.Equal(t, tt.want, cbc.NormalizeStrictCode(tt.code))
 		})
 	}
 }
@@ -354,43 +393,47 @@ func TestCode_Rules(t *testing.T) {
 			code: cbc.Code("123456789012345678901234567890AB"),
 		},
 		{
-			name:     "dot at start",
-			code:     cbc.Code(".B123"),
+			name: "dot at start",
+			code: cbc.Code(".B123"),
+		},
+		{
+			name: "dot at end",
+			code: cbc.Code("B123."),
+		},
+		{
+			name: "dash at start",
+			code: cbc.Code("-B123"),
+		},
+		{
+			name: "dash at end",
+			code: cbc.Code("B123-"),
+		},
+		{
+			name: "multiple symbols",
+			code: cbc.Code("AB/.CD"),
+		},
+		{
+			name: "multi-dash",
+			code: cbc.Code("AB--CD"),
+		},
+		{
+			name:     "leading space",
+			code:     cbc.Code(" B123"),
 			wantCode: "GOBL-CBC-CODE-02",
 		},
 		{
-			name:     "dot at end",
-			code:     cbc.Code("B123."),
+			name:     "trailing space",
+			code:     cbc.Code("B123 "),
 			wantCode: "GOBL-CBC-CODE-02",
 		},
 		{
-			name:     "dash at start",
-			code:     cbc.Code("-B123"),
-			wantCode: "GOBL-CBC-CODE-02",
-		},
-		{
-			name:     "dash at end",
-			code:     cbc.Code("B123-"),
-			wantCode: "GOBL-CBC-CODE-02",
-		},
-		{
-			name:     "multiple symbols",
-			code:     cbc.Code("AB/.CD"),
-			wantCode: "GOBL-CBC-CODE-02",
-		},
-		{
-			name:     "character return",
+			name:     "newline in middle",
 			code:     cbc.Code("AB\nCD"),
 			wantCode: "GOBL-CBC-CODE-02",
 		},
 		{
-			name:     "character return",
+			name:     "only newline",
 			code:     cbc.Code("\n"),
-			wantCode: "GOBL-CBC-CODE-02",
-		},
-		{
-			name:     "multi-dash",
-			code:     cbc.Code("AB--CD"),
 			wantCode: "GOBL-CBC-CODE-02",
 		},
 		{
@@ -468,44 +511,48 @@ func TestCode_Validate(t *testing.T) {
 			code: cbc.Code("123456789012345678901234567890AB"),
 		},
 		{
-			name:    "dot at start",
-			code:    cbc.Code(".B123"),
-			wantErr: "codes must only contain",
+			name: "dot at start",
+			code: cbc.Code(".B123"),
 		},
 		{
-			name:    "dot at end",
-			code:    cbc.Code("B123."),
-			wantErr: "codes must only contain",
+			name: "dot at end",
+			code: cbc.Code("B123."),
 		},
 		{
-			name:    "dash at start",
-			code:    cbc.Code("-B123"),
-			wantErr: "codes must only contain",
+			name: "dash at start",
+			code: cbc.Code("-B123"),
 		},
 		{
-			name:    "dash at end",
-			code:    cbc.Code("B123-"),
-			wantErr: "codes must only contain",
+			name: "dash at end",
+			code: cbc.Code("B123-"),
 		},
 		{
-			name:    "multiple symbols",
-			code:    cbc.Code("AB/.CD"),
-			wantErr: "codes must only contain",
+			name: "multiple symbols",
+			code: cbc.Code("AB/.CD"),
 		},
 		{
-			name:    "character return",
+			name: "multi-dash",
+			code: cbc.Code("AB--CD"),
+		},
+		{
+			name:    "leading space",
+			code:    cbc.Code(" B123"),
+			wantErr: "leading or trailing whitespace",
+		},
+		{
+			name:    "trailing space",
+			code:    cbc.Code("B123 "),
+			wantErr: "leading or trailing whitespace",
+		},
+		{
+			name:    "newline in middle",
 			code:    cbc.Code("AB\nCD"),
-			wantErr: "codes must only contain",
+			wantErr: "leading or trailing whitespace",
 		},
 		{
-			name:    "character return",
+			name:    "only newline",
 			code:    cbc.Code("\n"),
-			wantErr: "codes must only contain",
-		},
-		{
-			name:    "multi-dash",
-			code:    cbc.Code("AB--CD"),
-			wantErr: "codes must only contain",
+			wantErr: "leading or trailing whitespace",
 		},
 		{
 			name:    "too long",
