+++ data/rules/cbc.json
@@ -13,8 +13,8 @@
         },
         {
           "id": "GOBL-CBC-CODE-02",
-          "desc": "codes must only contain letters, numbers, and optionally separated by .-:/,_\u0026 or space",
-          "tests": "matches ^[A-ZÑa-z0-9]+([.\\-:/,_\u0026 ]?[A-ZÑa-z0-9]+)*$"
+          "desc": "codes must not have leading or trailing whitespace",
+          "tests": "matches ^\\S(.*\\S)?$"
         }
       ]
     },
