+++ org/inbox.go
@@ -39,37 +39,6 @@ type Inbox struct {
 	Email string `json:"email,omitempty" jsonschema:"title=Email"`
 }
 
-// Normalize will try to clean the inbox's data.
-func (i *Inbox) Normalize() {
-	if i == nil {
-		return
-	}
-	uuid.Normalize(&i.UUID)
-	code := i.Code.String()
-	if is.EmailFormat.Check(code) {
-		i.Email = code
-		i.Code = ""
-	} else if is.URL.Check(code) {
-		i.URL = code
-		i.Code = ""
-	}
-	i.Label = cbc.NormalizeString(i.Label)
-	i.Scheme = cbc.NormalizeUpperCode(i.Scheme)
-	i.Code = cbc.NormalizeCode(i.Code)
-
-	// Custom normalizations
-	switch i.Key {
-	case InboxKeyPeppol:
-		if i.Scheme == "" {
-			if len(i.Code) >= 5 && i.Code[4] == ':' {
-				numbers := i.Code[:4]
-				i.Scheme = numbers
-				i.Code = i.Code[5:]
-			}
-		}
-	}
-}
-
 func inboxRules() *rules.Set {
 	return rules.For(new(Inbox),
 		rules.Assert("01", "inbox requires a code, url, or email",
@@ -115,4 +84,34 @@ func AddInbox(in []*Inbox, i *Inbox) []*Inbox {
 		}
 	}
 	return append(in, i)
+}
+
+func normalizeInbox(i *Inbox) {
+	if i == nil {
+		return
+	}
+	uuid.Normalize(&i.UUID)
+	code := i.Code.String()
+	if is.EmailFormat.Check(code) {
+		i.Email = code
+		i.Code = ""
+	} else if is.URL.Check(code) {
+		i.URL = code
+		i.Code = ""
+	}
+	i.Label = cbc.NormalizeString(i.Label)
+	i.Scheme = cbc.NormalizeUpperCode(i.Scheme)
+	i.Code = cbc.NormalizeStrictCode(i.Code)
+
+	// Custom normalizations
+	switch i.Key {
+	case InboxKeyPeppol:
+		if i.Scheme == "" {
+			if len(i.Code) >= 5 && i.Code[4] == ':' {
+				numbers := i.Code[:4]
+				i.Scheme = numbers
+				i.Code = i.Code[5:]
+			}
+		}
+	}
 }
