+++ org/name_test.go
@@ -3,6 +3,7 @@ package org_test
 import (
 	"testing"
 
+	"github.com/invopop/gobl/norm"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/rules"
 	"github.com/stretchr/testify/assert"
@@ -13,7 +14,7 @@ func TestNameNormalize(t *testing.T) {
 	t.Run("nil", func(t *testing.T) {
 		var n *org.Name
 		assert.NotPanics(t, func() {
-			n.Normalize()
+			norm.Normalize(n)
 		})
 	})
 	n := &org.Name{
@@ -25,7 +26,7 @@ func TestNameNormalize(t *testing.T) {
 		Surname2: "  Smith  ",
 		Suffix:   "  Jr.  ",
 	}
-	n.Normalize()
+	norm.Normalize(n)
 	require.Equal(t, "john doe", n.Alias)
 	require.Equal(t, "Mr.", n.Prefix)
 	require.Equal(t, "John", n.Given)
