+++ org/party.go
@@ -62,3 +62,13 @@ func (Party) JSONSchemaExtend(js *jsonschema.Schema) {
 		},
 	}
 }
+
+func normalizeParty(p *Party) {
+	if p == nil {
+		return
+	}
+	uuid.Normalize(&p.UUID)
+	p.Label = cbc.NormalizeString(p.Label)
+	p.Name = cbc.NormalizeString(p.Name)
+	p.Alias = cbc.NormalizeString(p.Alias)
+}
