+++ org/registration_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/invopop/gobl/currency"
+	"github.com/invopop/gobl/norm"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/rules"
 	"github.com/stretchr/testify/assert"
@@ -23,7 +24,7 @@ func TestRegistrationNormalize_TrimsFieldsAndSetsUUID(t *testing.T) {
 	}
 
 	assert.NotPanics(t, func() {
-		r.Normalize()
+		norm.Normalize(r)
 	})
 
 	assert.Equal(t, "My Label", r.Label)
@@ -40,7 +41,7 @@ func TestRegistrationNormalize_TrimsFieldsAndSetsUUID(t *testing.T) {
 func TestRegistrationNormalize_OnNilReceiverDoesNotPanic(t *testing.T) {
 	var r *org.Registration
 	assert.NotPanics(t, func() {
-		r.Normalize()
+		norm.Normalize(r)
 	})
 }
 
@@ -50,7 +51,7 @@ func TestRegistrationValidate_ValidData(t *testing.T) {
 		Office:   "  Main Office  ",
 		Currency: currency.EUR,
 	}
-	r.Normalize()
+	norm.Normalize(r)
 
 	assert.NoError(t, rules.Validate(r))
 }
