+++ org/telephone_test.go
@@ -3,6 +3,7 @@ package org_test
 import (
 	"testing"
 
+	"github.com/invopop/gobl/norm"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/rules"
 	"github.com/stretchr/testify/assert"
@@ -12,14 +13,14 @@ func TestTelephoneNormalizeAndValidate(t *testing.T) {
 	t.Run("nil", func(t *testing.T) {
 		var tel *org.Telephone
 		assert.NotPanics(t, func() {
-			tel.Normalize()
+			norm.Normalize(tel)
 		})
 	})
 	t.Run("basic normalization", func(t *testing.T) {
 		tel := &org.Telephone{
 			Number: "  +123 456 7890  ",
 		}
-		tel.Normalize()
+		norm.Normalize(tel)
 		assert.Equal(t, "+123 456 7890", tel.Number)
 		assert.NoError(t, rules.Validate(tel))
 	})
@@ -28,7 +29,7 @@ func TestTelephoneNormalizeAndValidate(t *testing.T) {
 		tel := &org.Telephone{
 			Number: "   ",
 		}
-		tel.Normalize()
+		norm.Normalize(tel)
 		assert.Equal(t, "", tel.Number)
 		assert.ErrorContains(t, rules.Validate(tel), "telephone number is required")
 	})
@@ -37,7 +38,7 @@ func TestTelephoneNormalizeAndValidate(t *testing.T) {
 		tel := &org.Telephone{
 			Number: "+1 (123) 456-7890 ext. 123",
 		}
-		tel.Normalize()
+		norm.Normalize(tel)
 		assert.Equal(t, "+1 (123) 456-7890 ext. 123", tel.Number)
 		assert.NoError(t, rules.Validate(tel))
 	})
