+++ org/website.go
@@ -18,8 +18,15 @@ type Website struct {
 	URL string `json:"url" jsonschema:"title=URL,format=uri"`
 }
 
-// Normalize will try to remove any unnecessary whitespace from the website fields.
-func (w *Website) Normalize() {
+func websiteRules() *rules.Set {
+	return rules.For(new(Website),
+		rules.Field("url",
+			rules.Assert("01", "website URL is required and must be valid", is.Present, is.URL),
+		),
+	)
+}
+
+func normalizeWebsite(w *Website) {
 	if w == nil {
 		return
 	}
@@ -27,12 +34,4 @@ func (w *Website) Normalize() {
 	w.Label = cbc.NormalizeString(w.Label)
 	w.Title = cbc.NormalizeString(w.Title)
 	w.URL = cbc.NormalizeString(w.URL)
-}
-
-func websiteRules() *rules.Set {
-	return rules.For(new(Website),
-		rules.Field("url",
-			rules.Assert("01", "website URL is required and must be valid", is.Present, is.URL),
-		),
-	)
 }
