+++ pay/instructions_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/norm"
 	"github.com/invopop/gobl/pay"
 	"github.com/invopop/gobl/rules"
 	"github.com/invopop/gobl/tax"
@@ -21,14 +22,14 @@ func TestInstructionsNormalize(t *testing.T) {
 			"random": "",
 		}),
 	}
-	i.Normalize()
+	norm.Normalize(i)
 	assert.True(t, i.Ext.IsZero())
 	assert.Equal(t, "fooo", i.Ref.String())
 	assert.Equal(t, "Some random payment", i.Detail)
 
 	i = nil
 	assert.NotPanics(t, func() {
-		i.Normalize()
+		norm.Normalize(i)
 	})
 }
