+++ pay/norm.go
@@ -7,8 +7,8 @@ import (
 )
 
 // init registers the intrinsic normalizers for the pay package's types. The
-// norm engine handles recursion into nested values and the application of
-// regime/addon normalizers.
+// norm engine handles recursion, the global cleaning of codes and extensions,
+// and the application of regime/addon normalizers.
 func init() {
 	norm.Register("pay",
 		norm.For(normalizeRecord),
@@ -24,17 +24,14 @@ func normalizeRecord(r *Record) {
 	uuid.Normalize(&r.UUID)
 	r.Ref = cbc.NormalizeString(r.Ref)
 	r.Description = cbc.NormalizeString(r.Description)
-	r.Ext = r.Ext.Clean()
 }
 
 func normalizeInstructions(i *Instructions) {
 	if i == nil {
 		return
 	}
-	i.Ref = cbc.NormalizeCode(i.Ref)
 	i.Detail = cbc.NormalizeString(i.Detail)
 	i.Notes = cbc.NormalizeString(i.Notes)
-	i.Ext = i.Ext.Clean()
 }
 
 func normalizeTerms(t *Terms) {
@@ -42,5 +39,4 @@ func normalizeTerms(t *Terms) {
 		return
 	}
 	t.Notes = cbc.NormalizeString(t.Notes)
-	t.Ext = t.Ext.Clean()
 }
